1. Add --with-eigenvector to the configure command line (not documented yet). 2. Update verion number to today. 3. Fix places where SimpleVector.size() was assumed to be size_t.
build.sh: only use boost vector for build.sh test case.
cppad_eigen.hpp: reducde include so can use for CPPAD_TEST_VECTOR.
limits.hpp: adapt to use by cppad_eigen.hpp.
cppad_eigen.cpp: adapt to change in cppad_eigen.hpp.
pow.cpp: Fix mistake in previous autmatic edit if test vectors.

git-svn-id: https://projects.coin-or.org/svn/CppAD/trunk@2452 0e562018-b8fb-0310-a933-ecf073c7c197

+++ cppad/index_sort.hpp
@@ -144,11 +144,11 @@ void index_sort(const VectorKey& keys, VectorSize& ind)
 	typedef index_sort_element<Compare> element;
 
 	CPPAD_ASSERT_KNOWN(
-		keys.size() == ind.size(),
+		size_t(keys.size()) == size_t(ind.size()),
 		"index_sort: vector sizes do not match"
 	);
 
-	size_t size_work = keys.size();
+	size_t size_work = size_t(keys.size());
 	size_t size_out;
 	element* work = 
 		thread_alloc::create_array<element>(size_work, size_out);
