+++ cppad/local/dependent.hpp
@@ -190,12 +190,12 @@ void ADFun<Base>::Dependent(const ADvector &x, const ADvector &y)
 	);
 	ADTape<Base> *tape = AD<Base>::tape_ptr(x[0].tape_id_);
 	CPPAD_ASSERT_KNOWN(
-		tape->size_independent_ == x.size(),
+		tape->size_independent_ == size_t( x.size() ),
 		"Dependent: independent variable vector has been changed."
 	);
 # ifndef NDEBUG
 	size_t i, j;
-	for(j = 0; j < x.size(); j++)
+	for(j = 0; j < size_t(x.size()); j++)
 	{	CPPAD_ASSERT_KNOWN(
 		size_t(x[j].taddr_) == (j+1),
 		"ADFun<Base>: independent variable vector has been changed."
@@ -205,7 +205,7 @@ void ADFun<Base>::Dependent(const ADvector &x, const ADvector &y)
 		"ADFun<Base>: independent variable vector has been changed."
 		);
 	}
-	for(i = 0; i < y.size(); i++)
+	for(i = 0; i < size_t(y.size()); i++)
 	{	CPPAD_ASSERT_KNOWN(
 		CppAD::Parameter( y[i] ) | (y[i].tape_id_ == x[0].tape_id_) ,
 		"ADFun<Base>: dependent vector contains a variable for"
